Use IamHmacPasswordEncoder in app configuration
