+++ iam-login-service/src/main/java/it/infn/mw/iam/config/security/IamWebSecurityConfig.java
@@ -41,8 +41,8 @@ import org.springframework.security.config.annotation.web.configuration.EnableWe
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.crypto.password.NoOpPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.oauth2.provider.client.ClientCredentialsTokenEndpointFilter;
 import org.springframework.security.oauth2.provider.error.OAuth2AuthenticationEntryPoint;
 import org.springframework.security.oauth2.provider.expression.OAuth2WebSecurityExpressionHandler;
 import org.springframework.security.web.AuthenticationEntryPoint;
@@ -66,8 +66,8 @@ import it.infn.mw.iam.authn.HintAwareAuthenticationEntryPoint;
 import it.infn.mw.iam.authn.multi_factor_authentication.ExtendedAuthenticationFilter;
 import it.infn.mw.iam.authn.multi_factor_authentication.ExtendedHttpServletRequestFilter;
 import it.infn.mw.iam.authn.multi_factor_authentication.MultiFactorVerificationFilter;
-import it.infn.mw.iam.authn.oidc.OIDCAuthenticationProvider;
 import it.infn.mw.iam.authn.oidc.OIDCAuthenticationFilter;
+import it.infn.mw.iam.authn.oidc.OIDCAuthenticationProvider;
 import it.infn.mw.iam.authn.x509.IamX509AuthenticationProvider;
 import it.infn.mw.iam.authn.x509.IamX509AuthenticationUserDetailService;
 import it.infn.mw.iam.authn.x509.IamX509PreauthenticationProcessingFilter;
@@ -77,6 +77,7 @@ import it.infn.mw.iam.config.IamProperties.ExternalAuthAttributeSectionBehaviour
 import it.infn.mw.iam.config.IamProperties.RegistrationField;
 import it.infn.mw.iam.config.mfa.IamTotpMfaProperties;
 import it.infn.mw.iam.core.IamLocalAuthenticationProvider;
+import it.infn.mw.iam.core.client.IamHmacPasswordEncoder;
 import it.infn.mw.iam.core.oauth.TokenEndpointJwtClientAuthFilter;
 import it.infn.mw.iam.core.oidc.AuthorizationRequestFilter;
 import it.infn.mw.iam.persistence.repository.IamAccountRepository;
@@ -482,6 +483,9 @@ public class IamWebSecurityConfig {
     private UserDetailsService userDetailsService;
     private TokenEndpointJwtClientAuthFilter jwtClientAuthFilter;
 
+    @Value("${iam.client.secret-encoder-key}")
+    private String secretEncoderKey;
+
     public IntrospectEndpointAuthorizationConfig(
         OAuth2AuthenticationEntryPoint authenticationEntryPoint,
         @Qualifier("clientUserDetailsService") UserDetailsService userDetailsService,
@@ -496,7 +500,7 @@ public class IamWebSecurityConfig {
     protected void configure(final AuthenticationManagerBuilder auth) throws Exception {
 
       auth.userDetailsService(userDetailsService)
-        .passwordEncoder(NoOpPasswordEncoder.getInstance());
+        .passwordEncoder(new IamHmacPasswordEncoder(secretEncoderKey));
     }
 
     @Override
@@ -511,6 +515,56 @@ public class IamWebSecurityConfig {
         .httpBasic().authenticationEntryPoint(authenticationEntryPoint).and()
         .addFilterBefore(jwtClientAuthFilter, BasicAuthenticationFilter.class)
         .authorizeRequests().anyRequest().fullyAuthenticated();
+      // @formatter:on
+    }
+  }
+
+  @Configuration
+  @Order(16)
+  public static class RevokeEndpointAuthorizationConfig extends WebSecurityConfigurerAdapter {
+
+    @Autowired
+    private OAuth2AuthenticationEntryPoint authenticationEntryPoint;
+
+    @Autowired
+    @Qualifier("clientUserDetailsService")
+    private UserDetailsService userDetailsService;
+
+    @Value("${iam.client.secret-encoder-key}")
+    private String secretEncoderKey;
+
+    @Override
+    protected void configure(final AuthenticationManagerBuilder auth) throws Exception {
+
+      auth.userDetailsService(userDetailsService)
+        .passwordEncoder(new IamHmacPasswordEncoder(secretEncoderKey));
+    }
+
+    private ClientCredentialsTokenEndpointFilter clientCredentialsEndpointFilter()
+        throws Exception {
+
+      ClientCredentialsTokenEndpointFilter filter =
+          new ClientCredentialsTokenEndpointFilter("/revoke");
+      filter.setAuthenticationManager(authenticationManager());
+      return filter;
+    }
+
+    @Override
+    protected void configure(final HttpSecurity http) throws Exception {
+
+      // @formatter:off
+      http.antMatcher("/revoke**")
+        .httpBasic()
+          .authenticationEntryPoint(authenticationEntryPoint)
+        .and()
+          .addFilterBefore(clientCredentialsEndpointFilter(), BasicAuthenticationFilter.class)
+        .cors()
+        .and()
+        .exceptionHandling()
+          .authenticationEntryPoint(authenticationEntryPoint)
+        .and()
+          .csrf().disable()
+          .sessionManagement().sessionCreationPolicy(STATELESS);
       // @formatter:on
     }
   }
