+++ iam-login-service/src/main/java/it/infn/mw/iam/core/authn/TokenEndpointBasicAuthenticationProvider.java
@@ -19,17 +19,18 @@ import java.net.URLDecoder;
 import java.nio.charset.StandardCharsets;
 
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.authentication.AuthenticationServiceException;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.crypto.password.NoOpPasswordEncoder;
 import org.springframework.stereotype.Component;
 
 import it.infn.mw.iam.api.client.service.ClientService;
 import it.infn.mw.iam.core.client.ClientUserDetailsService;
+import it.infn.mw.iam.core.client.IamHmacPasswordEncoder;
 import it.infn.mw.iam.persistence.model.ClientAuthMethod;
 import it.infn.mw.iam.persistence.model.ClientDetailsEntity;
 
@@ -39,10 +40,13 @@ public class TokenEndpointBasicAuthenticationProvider extends DaoAuthenticationP
 
   private final ClientService clientService;
 
+  @Value("${iam.client.secret-encoder-key}")
+  private String secretEncoderKey;
+
   public TokenEndpointBasicAuthenticationProvider(
       @Qualifier("clientUserDetailsService") ClientUserDetailsService userDetailsService) {
 
-    this.setPasswordEncoder(NoOpPasswordEncoder.getInstance());
+    this.setPasswordEncoder(new IamHmacPasswordEncoder(secretEncoderKey));
     this.setUserDetailsService(userDetailsService);
     this.clientService = userDetailsService.getClientDetailsService();
   }
